% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wts_gen.R
\name{ebal_wts_simple}
\alias{ebal_wts_simple}
\title{Simple entropy balancing weights for causal generalization}
\usage{
ebal_wts_simple(x, target_moments = NULL, H_add_intercept = TRUE, delta)
}
\arguments{
\item{x}{A data matrix for the source sample. Each column represents
source sample covariate and each row represents an observation.}

\item{target_moments}{A vector of first moments of the target sample covariates
that needs to be balanced between source and target.}

\item{H_add_intercept}{`logical` whether to include 1 as intercept in
H covariates, default as TRUE.}

\item{delta}{A vector specifying the approximate balancing tolerance margin.
The vector has a total length of \code{2*ncol(x)}. If exact balancing, delta are all zeros.}
}
\value{
A list containing:
\item{w}{A vector of entropy balancing weights for causal generalization.}
\item{theta}{The dual parameters estimated from the optimization process.}
}
\description{
Compute the simple exact and approximate entropy balancing weights.
The resulting weights calibrate the whole x sample to the target moments (not
distinguishing treated and control), which is equivalent to an exponential
tilting calibration. If the tolerance margin \code{delta} is all 0,
it computes exact balancing weights. Otherwise, it computes approximate balancing weights.
}
\examples{
library(EBalGen)
set.seed(1)
n = 100
p = 5
x = runif(n * p)
x = matrix(4 * x - 2, n, p)
target_moments = c(0,0,0,0,0)
delta = numeric(5)
ebal_wts_simple(x, target_moments = target_moments, H_add_intercept = TRUE,delta)

## Approximate balancing weights requires installation of MOSEK for optimization.
delta = numeric(5)+0.1
if (requireNamespace("CVXR", quietly = TRUE)) {
  if (!("MOSEK" \%in\% CVXR::installed_solvers())) {
      stop("MOSEK solver is not installed. Please install MOSEK to use it with CVXR.")}
  ebal_wts_simple(x, target_moments = target_moments, H_add_intercept = TRUE,delta)

}
}
