% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPM_AB_Delta.R
\name{RPM_AB}
\alias{RPM_AB}
\title{Confidence interval for approximate balancing}
\usage{
RPM_AB(
  x,
  y,
  trt,
  H_vars,
  target_mean,
  target_sd,
  num_sim,
  H_add_intercept = TRUE,
  cluster = 1,
  set_seed = 111
)
}
\arguments{
\item{x}{A covariate data matrix for the source sample.}

\item{y}{A vector of the source sample response values.}

\item{trt}{A vector of 0, 1 or FALSE/TRUE of treatment assignment for the source sample.}

\item{H_vars}{A vector of numbers indexing which covariate in \code{x}
need to be balanced between source and target samples.}

\item{target_mean}{A vector of mean of the target sample covariates
that needs to be balanced between source and target.}

\item{target_sd}{A vector of standard deviation of the target sample covariates
that needs to be balanced between source and target.}

\item{num_sim}{`numeric` the number of simulation used for bootstrap.}

\item{H_add_intercept}{`logical` whether to include 1 as intercept in
H covariates, default as TRUE.}

\item{cluster}{Number of parallel running CPU cores, Default: 1}

\item{set_seed}{Random seed for simulation, Default: 111}
}
\value{
A list containing:
\describe{
 \item{mean_ATE}{The mean ATE over bootstrap.}
 \item{lb_ATE}{The lower bound of 95\% CI.}
 \item{ub_ATE}{The upper bound of 95\% CI.}
 \item{use_exact}{The number of times there no need for approximate balancing in the bootstrap}
 \item{n_success}{The number of feasible solutions in `num_sim` bootstrap.}
}
}
\description{
Compute the approximate balancing confidence interval (CI)
using resampling-based perturbation method proposed in the paper.
}
\examples{
library(EBalGen)
set.seed(1)
n = 100
p = 5
x = runif(n * p)
x = matrix(10 * x - 2, n, p)
y = rnorm(n*p)
trt = rbinom(n,1,0.5)
H_vars = c(1,2,3)
target_mean = c(2,2,2)
target_sd=c(1,1,1)
if (requireNamespace(c("dplyr","doRNG","rockchalk","resample","doParallel"),
  quietly = TRUE)) {

  RPM_AB(x,y,trt,H_vars, target_mean, target_sd, num_sim=50,
  H_add_intercept = TRUE,cluster=1, set_seed=111)

}
}
\references{
Chen, Y., Chen, G., & Yu, M. (2025+). Confidence interval construction for
causally generalized estimators with summary-level target data.
}
