% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPM_AB_Delta.R
\name{RPM_AB_delta}
\alias{RPM_AB_delta}
\title{Confidence interval for approximate balancing for a specified margin}
\usage{
RPM_AB_delta(
  x,
  y,
  trt,
  H_vars,
  target_mean,
  target_sd,
  num_sim,
  H_add_intercept = TRUE,
  delta,
  cluster = 1,
  set_seed = 111
)
}
\arguments{
\item{x}{A covariate data matrix for the source sample.}

\item{y}{Logical vector of treatment assignment for the source sample.}

\item{trt}{A vector of 0, 1 or FALSE/TRUE of treatment assignment for the source sample.}

\item{H_vars}{A vector of numbers indexing which covariate in \code{x}
need to be balanced between source and target samples.}

\item{target_mean}{A vector of mean of the target sample covariates
that needs to be balanced between source and target.}

\item{target_sd}{A vector of standard deviation of the target sample covariates
that needs to be balanced between source and target.}

\item{num_sim}{`numeric` the number of simulation used for bootstrap.}

\item{H_add_intercept}{`logical` whether to include 1 as intercept in
H covariates, default as TRUE.}

\item{delta}{A vector specifying the approximate balancing tolerance margin.
The vector has a total length of 2H+G, where H represents the number of
covariates balanced between the source (treatment and control) and the
target moments, and G represents the covariates balanced solely between
the source treatment and control groups. If exact balancing, delta are all zeros.}

\item{cluster}{Number of parallel running CPU cores, Default: 1}

\item{set_seed}{Random seed for simulation, Default: 111}
}
\value{
A list containing:
\describe{
 \item{mean_ATE}{The mean ATE over bootstrap.}
 \item{lb_ATE}{The lower bound of 95\% CI.}
 \item{ub_ATE}{The upper bound of 95\% CI.}
 \item{n_success}{The number of feasible solutions in `num_sim` bootstrap.}
}
}
\description{
Compute the approximate balancing confidence interval (CI) for a specified margin
using resampling-based perturbation method proposed in the paper.
}
\examples{
library(EBalGen)
set.seed(1)
n = 100
p = 5
x = runif(n * p)
x = matrix(6 * x - 2, n, p)
y = rnorm(n*p)
trt = rbinom(n,1,0.5)
H_vars = c(1,2,3)
target_mean = c(0,0,0)
target_sd=c(1,1,1)
delta = numeric(8)+0.1
if (requireNamespace(c("dplyr","doRNG","rockchalk","resample","doParallel"),
  quietly = TRUE)) {
  \dontrun{
  RPM_AB_delta(x,y,trt,H_vars, target_mean, target_sd, num_sim=10,
  H_add_intercept = TRUE,delta,cluster=1, set_seed=111)
}
}
}
\references{
Chen, Y., Chen, G., & Yu, M. (2025+). Confidence interval construction for
causally generalized estimators with summary-level target data.
}
