% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPM_CI.R
\name{RPM_CI}
\alias{RPM_CI}
\title{Confidence interval for exact balancing ATE}
\usage{
RPM_CI(
  x,
  y,
  trt,
  H_vars,
  target_mean,
  target_sd,
  num_sim,
  H_add_intercept = TRUE,
  cluster = 1,
  set_seed = 111
)
}
\arguments{
\item{x}{A data matrix for the source sample. Each column represents
source sample covariate and each row represents an observation.}

\item{y}{A vector of the source sample response values.}

\item{trt}{A vector of 0, 1 or FALSE/TRUE of treatment assignment for the source sample.}

\item{H_vars}{A vector of numbers indexing which covariate in `x`
need to be balanced between source and target samples.}

\item{target_mean}{A vector of mean of the target sample covariates
that needs to be balanced between source and target.}

\item{target_sd}{A vector of standard deviation of the target sample covariates
that needs to be balanced between source and target.}

\item{num_sim}{`numeric` the number of simulations used in bootstrap.}

\item{H_add_intercept}{`logical` whether to include 1 as intercept in
H covariates, default as TRUE.}

\item{cluster}{Number of parallel running CPU cores, Default: 1}

\item{set_seed}{Random seed for simulation, Default: 111}
}
\value{
A list containing:
\describe{
 \item{mean_ATE}{The mean ATE over bootstrap.}
 \item{lb_ATE}{The lower bound of 95\% CI.}
 \item{ub_ATE}{The upper bound of 95\% CI.}
 \item{n_success}{The number of feasible solutions in \code{num_sim} bootstrap.}
}
}
\description{
Compute the exact balancing ATE confidence interval (CI)
using resampling-based perturbation method proposed in the paper.
}
\examples{
library(EBalGen)
set.seed(1)
n = 100
p = 5
x = runif(n * p)
x = matrix(4 * x - 2, n, p)
y = rnorm(n*p)
trt = rbinom(n,1,0.5)
H_vars = c(1,2,3)
target_mean = c(0,0,0)
target_sd=c(1,1,1)
## Get CI for this generalized ATE
RPM_CI(x,y,trt,H_vars, target_mean, target_sd, num_sim=50,
H_add_intercept = TRUE,cluster=1, set_seed=111)

}
\references{
Chen, R., Chen, G., & Yu, M. (2023). Entropy balancing for causal
generalization with target sample summary information. Biometrics, 79(4)
}
